add env check; add logger; add exceptions

+++ src/events/events.controller.ts
@@ -4,6 +4,8 @@ import {
   Delete,
   Get,
   HttpCode,
+  Logger,
+  NotFoundException,
   Param,
   ParseUUIDPipe,
   Patch,
@@ -13,23 +15,31 @@ import {
 import { CreateEventDto } from 'src/dto/create-event.dto';
 import { UpdateEventDto } from 'src/dto/update-event.dto';
 import { EventsService } from './events.service';
+import { NotFoundError } from 'rxjs';
 
 @Controller('/events')
 export class EventsController {
+  private readonly logger = new Logger(EventsController.name);
+
   constructor(private readonly eventsService: EventsService) {}
   @Get()
   async findAll() {
-    return this.eventsService.findAll();
+    this.logger.log(`findAll route`);
+    const events = await this.eventsService.findAll();
+    this.logger.debug(`Found ${events.length} events`);
+    return events;
   }
   @Get(':id')
   async findOne(@Param('id', ParseUUIDPipe) id: string) {
-    return this.eventsService.findOne(id);
+    const event = await this.eventsService.findOne(id);
+    if (!event) {
+      throw new NotFoundException();
+    }
+    return event;
   }
 
   @Post()
-  async create(
-    @Body() input: CreateEventDto,
-  ) {
+  async create(@Body() input: CreateEventDto) {
     return this.eventsService.create(input);
   }
 
@@ -44,6 +54,10 @@ export class EventsController {
   @Delete(':id')
   @HttpCode(204)
   async remove(@Param('id', ParseUUIDPipe) id: string) {
+    const event = await this.eventsService.findOne(id);
+    if (!event) {
+      throw new NotFoundException();
+    }
     return this.eventsService.remove(id);
   }
 }
